revert type change

+++ sheeprl/algos/dreamer_v3/utils.py
@@ -82,7 +82,7 @@ def prepare_obs(
 ) -> Dict[str, Tensor]:
     torch_obs = {}
     for k, v in obs.items():
-        torch_obs[k] = torch.from_numpy(v.copy()).to(fabric.device)
+        torch_obs[k] = torch.from_numpy(v.copy()).to(fabric.device).float()
         if k in cnn_keys:
             torch_obs[k] = torch_obs[k].view(1, num_envs, -1, *v.shape[-2:]) / 255 - 0.5
         else:
